# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: مستندات API سرور احراز هویت Master OAuth
  description: |
    وب‌سرویس‌های REST و OAuth2 برای Authorization Server.

    ## احراز هویت
    بیشتر مسیرهای `/api/**` به توکن Bearer از `/oauth/token` نیاز دارند.

    ## جریان‌ها
    - **authorization_code** — لاگین مرورگر + کپچا + انتخاب scope
    - **password** — کلاینت‌های ماشینی (`api-client`)
    - **refresh_token** / **client_credentials** — مطابق تنظیمات هر کلاینت

    عملیات CRUD کاربران و Scope به دسترسی `write` نیاز دارد.
  version: 1.0.0
  contact:
    name: Master OAuth
...
servers:
  - url: http://localhost:8089
    description: محیط توسعه محلی
...
paths:
...
  /api/scopes:
    get:
      tags: [اسکوپ‌ها, عمومی]
      summary: فهرست اسکوپ‌ها
      operationId: listScopes
      parameters:
        - name: activeOnly
          in: query
          description: فقط اسکوپ‌های فعال
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: لیست اسکوپ‌ها
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScopeEntity'
    post:
      tags: [اسکوپ‌ها]
      summary: ایجاد اسکوپ
      operationId: createScope
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScopeRequest'
            example:
              name: reports
              displayName: گزارش‌ها
              description: دسترسی به API گزارش‌گیری
              active: true
      responses:
        '201':
          description: ایجاد شد
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeEntity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: نیاز به scope با نام write
        '409':
          description: نام اسکوپ تکراری است
  /api/scopes/{id}:
    get:
      tags: [اسکوپ‌ها, عمومی]
      summary: دریافت اسکوپ با شناسه
      operationId: getScope
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: جزئیات اسکوپ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeEntity'
        '404':
          description: یافت نشد
    put:
      tags: [اسکوپ‌ها]
      summary: به‌روزرسانی اسکوپ
      operationId: updateScope
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScopeRequest'
      responses:
        '200':
          description: به‌روزرسانی شد
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeEntity'
        '403':
          description: نیاز به scope با نام write
        '404':
          description: یافت نشد
    delete:
      tags: [اسکوپ‌ها]
      summary: حذف اسکوپ
      description: اسکوپ از همه کاربران نیز حذف می‌شود.
      operationId: deleteScope
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: حذف شد
        '403':
          description: نیاز به scope با نام write
        '404':
          description: یافت نشد
  /api/users:
    get:
      tags: [کاربران]
      summary: فهرست کاربران
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: لیست کاربران
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserEntity'
        '403':
          description: نیاز به scope با نام write
    post:
      tags: [کاربران]
      summary: ایجاد کاربر
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              username: charlie
              password: password
              email: charlie@example.com
              fullName: کاربر چارلی
              enabled: true
              scopes: [read, profile]
      responses:
        '201':
          description: ایجاد شد
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
        '403':
          description: نیاز به scope با نام write
        '409':
          description: نام کاربری تکراری است
  /api/users/{id}:
    get:
      tags: [کاربران]
      summary: دریافت کاربر با شناسه
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: جزئیات کاربر
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
        '404':
          description: یافت نشد
    put:
      tags: [کاربران]
      summary: به‌روزرسانی کاربر
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: به‌روزرسانی شد
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
        '404':
          description: یافت نشد
    delete:
      tags: [کاربران]
      summary: حذف کاربر
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: حذف شد
        '404':
          description: یافت نشد
  /api/users/{id}/scopes:
    put:
      tags: [کاربران]
      summary: جایگزینی اسکوپ‌های کاربر
      operationId: replaceUserScopes
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scopes]
              properties:
                scopes:
                  type: array
                  description: نام اسکوپ‌ها
                  items:
                    type: string
            example:
              scopes: [read, write, profile]
      responses:
        '200':
          description: کاربر به‌روز شد
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntity'
...
  /api/tokens/user/{username}:
    get:
      tags: [توکن‌ها]
      summary: فهرست توکن‌های یک کاربر (نیاز به write)
      operationId: listUserTokens
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          description: نام کاربری
          schema:
            type: string
      responses:
        '200':
          description: لیست توکن‌ها
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TokenRegistryView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: scope با نام write وجود ندارد
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: توکن دسترسی از `/oauth/token` (برای CRUD به scope با نام write نیاز است)
    basicAuth:
      type: http
      scheme: basic
      description: شناسه و رمز کلاینت OAuth
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: شناسه عددی رکورد
      schema:
        type: integer
        format: int64
  responses:
    Unauthorized:
      description: توکن موجود نیست، نامعتبر است یا ابطال شده
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OAuthError'
  schemas:
...
    ScopeEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: شناسه
        name:
          type: string
          description: نام اسکوپ
        displayName:
          type: string
          description: عنوان نمایشی
        description:
          type: string
          description: توضیح
        active:
          type: boolean
          description: فعال بودن
    CreateScopeRequest:
      type: object
      required: [name, displayName]
      properties:
        name:
          type: string
          description: نام یکتای اسکوپ
        displayName:
          type: string
          description: عنوان فارسی/نمایشی
        description:
          type: string
          description: توضیح
        active:
          type: boolean
          default: true
          description: فعال بودن
    UpdateScopeRequest:
      type: object
      properties:
        name:
          type: string
          description: نام اسکوپ
        displayName:
          type: string
          description: عنوان نمایشی
        description:
          type: string
          description: توضیح
        active:
          type: boolean
          description: فعال بودن
    UserEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: شناسه کاربر
        username:
          type: string
          description: نام کاربری
        email:
          type: string
          description: ایمیل
        fullName:
          type: string
          description: نام کامل
        enabled:
          type: boolean
          description: فعال بودن حساب
        accountNonExpired:
          type: boolean
          description: منقضی نبودن حساب
        accountNonLocked:
          type: boolean
          description: قفل نبودن حساب
        credentialsNonExpired:
          type: boolean
          description: منقضی نبودن اعتبارنامه
        createdAt:
          type: string
          format: date-time
          description: زمان ایجاد
        scopes:
          type: array
          description: اسکوپ‌های مجاز
          items:
            type: string
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: نام کاربری
        password:
          type: string
          format: password
          description: رمز عبور خام (در سرور hash می‌شود)
        email:
          type: string
          description: ایمیل
        fullName:
          type: string
          description: نام کامل
        enabled:
          type: boolean
          default: true
          description: فعال بودن
        scopes:
          type: array
          description: لیست نام اسکوپ‌ها
          items:
            type: string
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          description: نام کاربری
        password:
          type: string
          format: password
          description: در صورت ارسال، رمز دوباره encode می‌شود
        email:
          type: string
          description: ایمیل
        fullName:
          type: string
          description: نام کامل
        enabled:
          type: boolean
          description: فعال بودن
        accountNonLocked:
          type: boolean
          description: قفل نبودن
        accountNonExpired:
          type: boolean
          description: منقضی نبودن حساب
        credentialsNonExpired:
          type: boolean
          description: منقضی نبودن اعتبارنامه
        scopes:
          type: array
          description: اسکوپ‌ها
          items:
            type: string
...
    OAuthError:
      type: object
      properties:
        error:
          type: string
          description: کد خطا
        error_description:
          type: string
          description: توضیح خطا
...
    TokenRegistryView:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: شناسه رجیستری
        jti:
          type: string
          description: شناسه توکن
        tokenType:
          type: string
          description: نوع توکن
        username:
          type: string
          description: نام کاربری
        clientId:
          type: string
          description: شناسه کلاینت
        scopes:
          type: string
          description: اسکوپ‌ها
        issuedAt:
          type: string
          format: date-time
          description: زمان صدور
        expiresAt:
          type: string
          format: date-time
          description: زمان انقضا
        revoked:
          type: boolean
          description: ابطال شده؟
        revokedAt:
          type: string
          format: date-time
          nullable: true
          description: زمان ابطال
        active:
          type: boolean
          description: فعال بودن
...